Medicine dao pta

+++ src/components/medicine/MedicineList.js
@@ -21,32 +21,40 @@ function MedicineList() {
 
 // ------------------------------------------- delete -------------------------------------------------
     const handleDelete = async () => {
-        swal.fire({
-            title: "Bạn có muốn xoá sản phẩm này khỏi giỏ hàng?",
-            text: selectMedicine.name,
-            icon: "warning",
-            showCancelButton: true,
-            confirmButtonColor: "#3085D6",
-            cancelButtonColor: "#d33",
-            confirmButtonText: "Đồng ý!",
-        })
-            .then(async (willDelete) => {
-                if (willDelete.isConfirmed) {
-                    await medicineService.deleteMedicine(selectMedicine.id);
-                    swal.fire("Xoá sản phẩm thành công!", "", "success");
-                } else {
-                    swal.fire({
-                        icon: 'error',
-                        title: 'Rất tiếc...',
-                        text: 'Xóa thất bại!'
-                    });
-                    setSelectMedicine({
-                        id:null,
-                        name: ''
-                    })
-                }
-                await getListSearchMedicine(searchInMedicine, searchInput, page, limit);
-            });
+        if (selectMedicine.id == null) {
+            swal.fire({
+                icon: "error",
+                title: "Rất tiếc...",
+                text: "Vui lòng chọn khách hàng trước khi thực hiện thao tác này!",
+            })
+        } else {
+            swal.fire({
+                title: "Bạn có muốn xoá sản phẩm này khỏi giỏ hàng?",
+                text: selectMedicine.name,
+                icon: "warning",
+                showCancelButton: true,
+                confirmButtonColor: "#3085D6",
+                cancelButtonColor: "#d33",
+                confirmButtonText: "Đồng ý!",
+            })
+                .then(async (willDelete) => {
+                    if (willDelete.isConfirmed) {
+                        await medicineService.deleteMedicine(selectMedicine.id);
+                        swal.fire("Xoá sản phẩm thành công!", "", "success");
+                        setSelectMedicine({
+                            id: null,
+                            name: ''
+                        })
+                    } else {
+                        swal.fire({
+                            icon: 'error',
+                            title: 'Rất tiếc...',
+                            text: 'Xóa thất bại!'
+                        });
+                    }
+                    await getListSearchMedicine(searchInMedicine, searchInput, page, limit);
+                });
+        }
     };
 // ---------------------------------------- Get list ---------------------------------------------
 
@@ -70,8 +78,8 @@ function MedicineList() {
 // ----------------------------------------- Search ---------------------------------------
     const getListSearchMedicine = async (searchInMedicine, searchInput, page, limit) => {
         const result = await medicineService.searchMedicine(searchInMedicine, searchInput, page, limit);
-            setMedicineList(result?.content);
-            setTotalPage(result?.totalPages);
+        setMedicineList(result?.content);
+        setTotalPage(result?.totalPages);
     }
 // select child
     const handleShowCondition = () => {
@@ -104,7 +112,7 @@ function MedicineList() {
     }
 
     useEffect(() => {
-        getListSearchMedicine(searchInMedicine,searchInput, page, limit)
+        getListSearchMedicine(searchInMedicine, searchInput, page, limit)
     }, [searchInput, page, limit])
 
     if (!medicineList) {
@@ -113,8 +121,8 @@ function MedicineList() {
     return (
         <>
             <div className="container">
-                <div className="row header" style={{textAlign: 'center', color: '#0D6EFD'}}>
-                    <h1 className="mt-4 mb-3">DANH SÁCH THUỐC</h1>
+                <div className="row header" >
+                    <h1 className="mt-4 mb-3" style={{textAlign: 'center', color: '#0D6EFD'}}>DANH SÁCH THUỐC</h1>
                 </div>
                 <div className="row row-function" style={{display: 'flex'}}>
                     <div className="col-9 col-search d-flex align-items-center justify-content-start gap-3">
@@ -207,8 +215,14 @@ function MedicineList() {
                                 {
                                     medicineList.map((item, index) => (
                                         <tr key={index} id={index} onClick={() => {
-                                            setSelectMedicine({id: item.id, name: item?.name});
-                                        }} style={(selectMedicine.id === item?.id) ? {background: 'red'} : {}}>
+                                            if (selectMedicine === null || selectMedicine.id !== item.id) {
+                                                setSelectMedicine({id: item.id, name: item?.name});
+                                            } else if (selectMedicine.id === item.id) {
+                                                setSelectMedicine({id: null, name: ""});
+                                            }
+
+                                        }}
+                                            style={(selectMedicine.id === item?.id) ? {background: 'rgba(252, 245, 76, 0.73)'} : {}}>
                                             <td className="px-3 py-3 border-b border-gray-200 text-sm">{index + 1}</td>
                                             <td className="px-3 py-3 border-b border-gray-200 text-sm">{item.code}</td>
                                             <td className="px-3 py-3 border-b border-gray-200 text-sm">{item.kindOfMedicineName}</td>
